# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by id
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /courses/{courseId}/assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments in a course (includes current user's submission status)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create an assignment (teacher)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{assignmentId}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get an assignment by id
      parameters:
        - $ref: "#/components/parameters/AssignmentId"
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update an assignment and its targeted audience (teacher only)
      parameters:
        - $ref: "#/components/parameters/AssignmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentUpdate"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "403":
          description: Not the course teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: integer
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Course:
      type: object
      required: [id, title, teacherId, inviteCode, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        teacherId:
          type: string
        teacherName:
          type: ["string", "null"]
        inviteCode:
          type: string
        isActive:
          type: boolean
        studentCount:
          type: integer
        assignmentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    GroupMemberInfo:
      type: object
      required: [id, isLeader]
      properties:
        id:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isLeader:
          type: boolean
    CourseGroup:
      type: object
      required: [id, courseId, name, members, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        name:
          type: string
        leaderId:
          type: ["string", "null"]
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMemberInfo"
        createdAt:
          type: string
          format: date-time
...
    Assignment:
      type: object
      required: [id, courseId, title, allowedTypes, maxScore, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        allowedTypes:
          type: array
          items:
            type: string
            enum: [text, file, link, video, audio]
        maxScore:
          type: integer
        assignmentType:
          type: string
          enum: [individual, group]
        leaderOnlySubmit:
          type: boolean
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionFile"
        createdAt:
          type: string
          format: date-time
        targetGroupIds:
          type: ["array", "null"]
          description: Teacher-only. Group ids a group assignment targets.
          items:
            type: integer
        targetGroups:
          type: ["array", "null"]
          description: Teacher-only. Details for targeted groups.
          items:
            $ref: "#/components/schemas/CourseGroup"
        myGroup:
          description: Student-only on group assignments. The viewer's targeted group.
          oneOf:
            - $ref: "#/components/schemas/CourseGroup"
            - type: "null"
        submissionCount:
          type: integer
        mySubmissionId:
          type: ["integer", "null"]
        mySubmissionStatus:
          type: ["string", "null"]
          enum: [submitted, graded, null]
        targetStudentIds:
          type: ["array", "null"]
          description: Teacher-only. Student ids the assignment is targeted at; null/empty means all students.
          items:
            type: string
        targetStudents:
          type: ["array", "null"]
          description: Teacher-only. Details for targeted students.
          items:
            type: object
            required: [id]
            properties:
              id:
                type: string
              name:
                type: ["string", "null"]
              email:
                type: ["string", "null"]
    AssignmentInput:
      type: object
      required: [title, allowedTypes]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        dueDate:
          type: string
          format: date-time
        allowedTypes:
          type: array
          minItems: 1
          items:
            type: string
            enum: [text, file, link, video, audio]
        maxScore:
          type: integer
          minimum: 1
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionFile"
        targetStudentIds:
          type: array
          description: Optional. Enrolled student ids to target; omit or empty for all students.
          items:
            type: string
        assignmentType:
          type: string
          enum: [individual, group]
          description: Defaults to individual.
        targetGroupIds:
          type: array
          description: Required (non-empty) when assignmentType is group.
          items:
            type: integer
        leaderOnlySubmit:
          type: boolean
          description: Group assignments only. When true only the group leader may submit.
    AssignmentUpdate:
      type: object
      description: Partial update. Omitted fields are left unchanged. targetStudentIds, when present, replaces the audience (empty array = all students).
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        allowedTypes:
          type: array
          minItems: 1
          items:
            type: string
            enum: [text, file, link, video, audio]
        maxScore:
          type: integer
          minimum: 1
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionFile"
        targetStudentIds:
          type: array
          description: Replaces the targeted audience. Empty array = all students.
          items:
            type: string
...
    SubmissionFile:
      type: object
      required: [path, name]
      properties:
        path:
          type: string
        name:
          type: string
...